fix Bug #70049: the bug is caused by when replace column to check trap, it only add new columns to checktrap but do not remove replace column, so it will check wrong trap for column replace action.
in this case, it using two columns from two tables of n:1:n, and then add the third table column to replace the second table column, so there will only two tables for new binding to checktrap. So it should not show warnig for trap. chasm trap should only occur for the three tables.