Bug #70064
In Role, both Administrator and Organization Administrator are global roles.   Therefore, the org values of both are null, which leads to the failure of applying the permissions.   This issue is caused by the modification of bug 69738.   Bug 69738 only consider the cases where identityType is equal to Identity.USER.   Thus, it is only necessary to check other conditions that are not user type.