Bug 70061 (#26)
* fix Bug #70061:
the bug is caused by when using calc field as measure value in selection list, the source table can not find the column because calc field is saved in vs, so get calc from vs and add it to check trap.

* fix Bug #70061:
the bug is caused by when using calc field as measure value in selection list, the source table can not find the column because calc field is saved in vs, so get calc from vs and add it to check trap.

---------

Co-authored-by: catherineli <catherineli@inetsoft.com>