Merge forward Bug #69919 (#36)
Prevent bookmarks from getting lost. Make changes to the VSBookmark instance from the storage as it could have changes from other inetsoft server instances or other runtime viewsheets. Add a distributed lock around this area of the code to prevent race conditions.
Debounce the code to update last accessed time of a bookmark and do it in a separate thread.