Bug #70064
This bug occurred because when modifying bug69738, the fact that site admin and org admin are global was not taken into account.The global organization is null, so according to the modified logic, the permission setting for org admin will be lost. Therefore, it is only necessary to reconsider the situation where orgid is null.