Bug 70064 (#37)
* Bug #70064

In Role, both Administrator and Organization Administrator are global roles.   Therefore, the org values of both are null, which leads to the failure of applying the permissions.   This issue is caused by the modification of bug 69738.   Bug 69738 only consider the cases where identityType is equal to Identity.USER.   Thus, it is only necessary to check other conditions that are not user type.

* Bug #70064

This bug occurred because when modifying bug69738, the fact that site admin and org admin are global was not taken into account.The global organization is null, so according to the modified logic, the permission setting for org admin will be lost. Therefore, it is only necessary to reconsider the situation where orgid is null.

---------

Co-authored-by: ramontzhao <ramontzhao@inetsoft.com>