Bug #70112
The bug occurred because the 'Host Organization Global Repository' was added, and the path for this entry is 'Host Organization Global Repository'.  In the code, it only checks if the path is "/".  Since 'Host Organization Global Repository' was added, the 'Repository' folder appears as a sibling directory and is displayed as well.  Therefore, the path in the search results becomes an empty string ("").  As a result, the empty string should also be excluded from the check.