Bug #70256  Created "datasource.cloudError" property to allow custom error messages for cloud data sources connection tests. (#186)
Fixes the error message handling when testing a data source connection.
Also has a new message specifically for 504 errors, and an additional custom message for cloud hosted data sources.
This message is configured by setting the "datasource.cloudError" property.
The datasources are specified as "cloud-hosted" by setting the field in their listing class.