fix Bug #70329: support select time zone for task option dates.
When select date, it will create one date object in client with current local time zone, but it should set model to server with select time zone in model. So we should convert the date to right timestamp.

Such as: in local china local, we select 2025-01-02 00:00:00, it will create date 2025-01-02 00:00:00 UTC+8, and after date.getTime, it will get date 2025-01-01 16:00:00 UTC. It will be wrong.

In fact, if we select date 2025-01-02 00:00:00(2025-01-02 00:00:00 UTC+8) and select time zone to UTC0, we want timestamp to 2025-01-02 00:00:00 UTC0, so we should +8 hours for timestamp to get right time we want.

However, when populate date time to date picker, we should change time to local time zone to shwo right date. so that we can keep date select not changed after reopen pane.

In server, we should also check date is between start date and end date according to the select time zone not server time zone. Also because user date have its own time zone.