bug-70384
This bug is a special case.  When security is turned off and there is no login, the user is null when creating a VS in the composer, for example, 8^VIEWSHEET^_NULL_^Untitled-4^127.0.0.1~.  As a result, during the traversal in the em, there is an admin user and an anonymous user, both of which are not null.  This causes the autosavefile to be added twice.  Therefore, deduplication in the map can resolve this issue.