Bug #70094, store the last status in the data source. (#244)
Check all the data sources first before saving the data sources to the storage so that asset change events can be debounced.
Increased the debounce time for asset change to 2 seconds and set preserveSequence to false as otherwise the debouncer is almost useless there.
Don't fire a modified event when updating the status.