bug-70323
For example, if the folder name is "b$ aa @ & .  #{} 中文_()!,.*?", the presence of a comma (,) causes an issue when passing the parameter to the server.  This is because, by default, @RequestParam treats commas as delimiters, splitting the single string into multiple array elements.
As a result, when applying a filter, the comparison is made between "b$ aa @ & .  #{} 中文_()!" and "b$ aa @ & .  #{} 中文_()!,.*?", causing a mismatch and failure to find the folder.
To resolve this, the frontend should encode the parameter before sending it, and the backend should decode it upon receiving it to restore the original folder name.