bug-70552
The bug occurs because the icon was not retained from the backend when saving.  Instead, the frontend uniformly set the icon to "database-icon" for items with the type "DATA_SOURCE," leading to the issue.  To fix this, the backend should retrieve the corresponding icon based on the typename of the datasource and save it accordingly.