fix Bug #70599： when copy organiztion or ranme organization, it will send task add request to add new tasks to scheduleclient service. So it will get new info to run tasks.
when rename user, the user is used for task info(email info for backup action), after migate task, should add task to schedule client so there is no need to restart schedule client to run old tasks.

copy organization and rename user name in different method to migate, so will not fix it duplicate.

Bug #70567 is the same issue.