Bug-70679
This bug occurs because after making multiple selections on the pane, only the path of the last selected node was passed to the tree. Since the tree on the left and the pane on the right are bound to each other, after the tree receives the path of the last selected node, it selects the corresponding node in the tree, which is then passed back to the pane. As a result, the pane, which had multiple nodes selected, now only highlights the last selected node. This creates a closed loop. To fix it, simply add a new selectNode() method to separate the operations on the tree nodes from the operations on the pane nodes.