Bug #70780
Add moveOrgOrgUpdatedList, and consider the case where, when the retrieved permission set has a size of 1, it might still belong to the original organization.
When isRemovePermission is true, it is necessary to reset the old permission back to oldResource to ensure that the old permission does not contain any permissions for newResource.