Merge forward Bug #70746 (#768)
Update the last modified time of the logical model when the dependencies are updated. This ensures that the resulting blob digest also changes. Without this, there’s a risk that a previously seen digest could reappear, leading to a race condition where the blob is being removed and added at the same time.