Bug #70780
The organization should default to the current organization. For example, if the admin logs in and the current organization is org0, then org0 should be selected. There are also cases where allUserGrants > 1 and currentOrgUserGrants is null, which should follow the same logic as when allUserGrants == 1. Therefore, the condition allUserGrants.size() > 0 should be modified accordingly.