bug-71020
During the table rename process, since ouser is an IdentityID object and Tool.getValue(user) is a string (e.g., "admin~;~host-org"), the comparison always returns false. As a result, the logic to update the path in the document was not executed, which caused the current issue. Converting ouser to a string before comparison would resolve this problem.