bug-71028
Because entity. type = 1024 and node.data.type = 1025, reopening the page allows adding the same datasource again.  Bitwise comparison shows that both include DATA_SOURCE, but 1025 is actually 1024 | 1, meaning it also includes FOLDER.
Therefore, even though they refer to the same resource, the types are not exactly equal, which leads to them being treated as different objects.
Using a "bitwise contains" check instead of "strict equality"—to be consistent with the service—can resolve this issue.