Bug-71066
The bug occurs because the orgId was not passed when calling getPermissions, which caused orgId to be null during getResourceKey.  As a result, the key was generated using the default organization host-org, resulting in:
SECURITY_ORGANIZATION:host-org:organization0~;~organization0.
The correct key should be:
SECURITY_ORGANIZATION:organization0:organization0~;~organization0.
Adding the orgId parameter will resolve the issue.