Merge forward changes for Feature 71078 - viewer web component (#1001)
* Feature #71078
Add viewer web component.
Refactor the chart web component logic so it can be reused for the viewer.

* Feature #71078
Add support for a global loading indicator to be able to show only one loading icon in the viewer web component

* Feature #71078
Fix issues with images not showing when using <inetsoft-base>.
Hide the edit button in charts and tables in the viewer web component.

* Feature #71078
Fix path