bug-71204
The bug occurs because the event listener did not capture the event, resulting in the previously created action not being closed. The reason it wasn't captured is that this.renderer.listen("document", "mousedown", (e) => this.documentMousedown(e)) binds the listener in the bubbling phase by default. This means that if event.stopPropagation() is called elsewhere on the page, it can prevent the event from reaching the listener. Therefore, in the formula-editor-dialog, the blockMouse directive should be set to false to avoid stopping event propagation, which will resolve the issue.