Fix image display issues related to Bug #71924 changes (#1848)
* Bug #71924
Handle alignment issues with images generated on the server side such as when shadow is enabled.
Add more properties to image info so that the image is regenerated once the property is changed.
For dynamic image always send the response back as a png.

* Bug #72142
Handle drawing a shadow when background is set. Don't fill with background color on the image where shadow will be applied.
Size the svg correctly when rendered on the server to match the browser.

* Bug #72159
Fix svg image alignment in exports