Merge forward Bug #72270 (#1956)
Add a script function to refresh a dashboard.
Make sure that data in script also uses fresh data on refresh. Use the touch timestamp of the viewsheet sandbox when available such as when executing script in viewsheets.
Don't show the loading mask when using the script refresh function so that it doesn't appear as if the entire viewsheet is loading.