Bug 72438: Apply Button Enable Logic Fix (#2048)
* Fix Bug #72438:
Since this component directly assigns values based on the model, the oldModel should be reassigned after clicking "Apply". Otherwise, the oldModel will remain the initially opened model, which affects the disabling of the "Apply" button.

* Fix Bug #72438:
Since this component directly assigns values based on the model, the oldModel should be reassigned after clicking "Apply". Otherwise, the oldModel will remain the initially opened model, which affects the disabling of the "Apply" button.