Bug #72429: the bug is only occor when first time open server and server have no security and no dashboards. (only occur for server nodes config)
when first time load registry, the repository is null, it will not add changeListener for dataspace change. So after first node to add dashboard, the second node will not dispatch changeListener and will not clear cache to reload dashboard from data space.

in fact, we should add change listener for every registry when load registry.(because registry will be changed by other nodes)