fix Bug #72265: (#1964)
The tool.call used to call methods for db drivers, in old verion, it will cause classNotfound exception so it will get class from plugin. But in stylebi, we will get theses class from default classloader, but it will have methodAccestor for the methods, so it can not invoke method rightly. If catch exception, we should check if obj's class load is plugin class loader, if yes, should get method from plugin and then to invke it