Fix Bug #72605: (#2183)
* Fix Bug #72605:
Research has found that when using a map for get in Ignite, duplicate objects are only written once. Therefore, if loginUser and userId are the same, only one entry will be stored, which results in failing to retrieve the value. Moreover, since loginUser is not a core part of getUser(), it does not need to be added to usersMap.

* coding review

* coding review

* Fix Bug #72605:
should set login user name with loginAsUser but not remove it.