Bug #72455 (#2205)
Make sure that the session is released on sso logout and redirected properly when using the portal.logout.url property.
Fix session invalidation issues due to client info mismatch when SSO is enabled.
Ensure that the client info in the Principal object matches the users map in SecurityEngine to prevent sessions from being invalidated immediately and to allow proper session cleanup.