Bug #72633: Should save image and load it from runtime viewsheet. (#2185)
* Bug #72633:
the bug is caused by when get runtime viewsheet from RuntimeCache, it will not save and load ImageHashService, so its image will be lost. Should save image and load it from runtime viewsheet.

* Update ImageHashService.java

coding review, should using private for the map

* Update WorksheetEngine.java

update puRuntimeVS to putRuntimeSheet to use for vs and ws

* Update WorksheetService.java

update puRuntimeVS to putRuntimeSheet to use for vs and ws

* Update WorksheetEngine.java

update puRuntimeVS to putRuntimeSheet to use for vs and ws

* Update ComposerViewsheetService.java

update puRuntimeVS to putRuntimeSheet to use for vs and ws

* Update WorksheetEngine.java

for amap, should use putSheet to add sheet

* Update ImageHashService.java

check nulll for string type attributes