Bug #73056 (#2402)
Replaced InheritableThreadLocal with regular ThreadLocal to prevent principal objects from lingering in threads and being used incorrectly. Identified the places where new threads are spawned to determine where the context principal should be explicitly passed. Avoided setting a principal in GroupedThread unless required to prevent unknown principals from persisting for the lifetime of the thread, especially in thread pools.
Prevented unnecessary writes of the runtime sheet to the cache when being checked for existence in ViewsheetService.checkVSExisted().