Bug #73040 (#2431)
Fix the logic for updating sandboxes which ended up placing an embedded VS inside of a root sandbox.
Don't update the last modified time of embedded viewsheets after parsing. This caused the state to not be parsed in a deeply embedded VS because of the recursive nature of the update call. It doesn't seem to have an effect for Bug #25028 for which this change was added.